move service hooks to service
