--- src/services/0_ExternToolsMS/etherpad/hooks/index.js
@@ -1,11 +1,0 @@
-const padHooks = require('./Pad');
-const sessionHooks = require('./Session');
-const groupHooks = require('./Group');
-const authorHooks = require('./Author');
-
-module.exports = {
-	padHooks,
-	sessionHooks,
-	groupHooks,
-	authorHooks,
-};
