--- src/services/0_ExternToolsMS/etherpad/services/index.js
@@ -1,11 +1,0 @@
-const Pad = require('./Pad');
-const Session = require('./Session');
-const Group = require('./Group');
-const Author = require('./Author');
-
-module.exports = {
-	Pad,
-	Session,
-	Group,
-	Author,
-};
