--- src/services/0_ExternToolsMS/nexboard/services/Projects.js
@@ -1,29 +1,0 @@
-const nexboardClient = require('../../logic/NexboardClient');
-
-class Project {
-	constructor(options) {
-		this.options = options || {};
-		this.docs = {};
-	}
-
-	get(id, params) {
-		return nexboardClient.getProject(id);
-	}
-
-	find(params) {
-		return nexboardClient.getProjectsIds();
-	}
-
-	create({
-		title = 'Neues Nexboard Projekt',
-		description = 'Hier werden alle Nexboards für diese Lerneinheit gesammelt',
-	}, params) {
-		return nexboardClient.createProject(title, description);
-	}
-
-	setup(app) {
-		this.app = app;
-	}
-}
-
-module.exports = Project;
