--- src/services/0_ExternToolsMS/nexboard/services/index.js
@@ -1,7 +1,0 @@
-const Project = require('./Projects');
-const Board = require('./Boards');
-
-module.exports = {
-	Project,
-	Board,
-};
